Saves the salt as an 8-bit length value in HBase instead of short which translates to 4 bytes.

+++ hbase-pal/src/main/java/org/gooseman/hbase/HBase.java
@@ -24,5 +24,5 @@ public @interface HBase {
      * A value greater than zero will enable the internal key salting algorithm
      * @return
      */
-    short salt() default 0;
+    byte salt() default 0;
 }
